Replacing ui.py with rp_visualization.py
